0-hello-world_prompt.c
How It Helps:
This code introduces user input through the scanf function. It prompts the user to enter their name, which is then used in a greeting message. This version allows interaction with the program, teaching how to handle user input in C.

Practical Use:
Getting user input (in this case, a name) and using it for personalized output.

Learning how to use scanf to handle data provided by the user.

Useful for building interactive applications that need to take user input and respond accordingly.